% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveActivity.R
\name{archiveActivity}
\alias{archiveActivity}
\title{Archive an Activity}
\usage{
archiveActivity(activityId = NULL, token)
}
\arguments{
\item{activityId}{ID of an Activity, eg. 123.}

\item{token}{Token obtained with \code{\link{signIn}}.}
}
\value{
Nothing is returned, just a message if the activity has been archived.
}
\description{
Archive an Activity
}
\examples{
token <- "123456789"
archiveActivity(activityId = NULL token, as_df = TRUE)
}
