% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignActivity.R
\name{assignActivity}
\alias{assignActivity}
\title{Assign an Activity to Device Side}
\usage{
assignActivity(activityId = NULL, deviceSide = NULL, token, as_df = TRUE)
}
\arguments{
\item{activityId}{ID of an Activity, eg. 123.}

\item{deviceSide}{Side of an active Device.}

\item{token}{Token obtained with \code{\link{signIn}}.}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with id, name, color, integration and deviceSide.
}
\description{
With this endpoint you can assign an Activity to any Side of your active Device.
}
\examples{
token <- "123456789"
assignActivity(activityId = "57302", deviceSide = 8, token, as_df = TRUE)
}
