% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_mins_to_hm}
\alias{convert_mins_to_hm}
\title{Convert minutes to H:M format}
\usage{
convert_mins_to_hm(timeDiff)
}
\arguments{
\item{timeDiff}{Numeric}
}
\value{
A string containing the converted numeric into the format %H:%M
}
\description{
Convert minutes to H:M format
}
\examples{
convert_mins_to_hm(timeDiff = 69.9833)
}
