% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_to_posix}
\alias{convert_to_posix}
\title{Convert Timeular timestamp to POSIX}
\usage{
convert_to_posix(timeular_time, tz = "CET")
}
\arguments{
\item{timeular_time}{Timestamp from Timeular}

\item{tz}{Desired timezone}
}
\value{
A POSIX object with desired timezone
}
\description{
Convert Timeular timestamp to POSIX
}
\examples{
convert_to_posix(timeular_time = "2017-09-18T14:32:26.960", tz = "CET")
}
