% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentTracking.R
\name{currentTracking}
\alias{currentTracking}
\title{Show current tracking}
\usage{
currentTracking(token, as_df = TRUE, tz = "CET")
}
\arguments{
\item{token}{Token obtained with \code{\link{signIn}}}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}

\item{tz}{Desired timezone}
}
\value{
A data frame or list with id, name, color, integration, started at and note of current tracking.
}
\description{
Show current tracking
}
\examples{
token <- "123456789"
currentTracking(token, as_df = TRUE, tz = "CET")
}
