% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperSignInRequest.R
\name{DeveloperSignInRequest}
\alias{DeveloperSignInRequest}
\title{Obtain Access token}
\usage{
DeveloperSignInRequest(apiKey, apiSecret)
}
\arguments{
\item{apiKey}{Your API key}

\item{apiSecret}{Your secret API key}
}
\value{
An Access Token required to access secured endpoints
}
\description{
With this endpoint you can obtain Access Token required to access secured endpoints.
To do so, you have to provide API Key & API Secret.
They can be generated on the profile website (https://profile.timeular.com/#/login).
}
\examples{
apiKey <- "ABCDefgh1234="
apiSecret <- "EFGHijkl5678="
DeveloperSignInRequest(apiKey, apiSecret)
}
