% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchKey.R
\name{fetchKey}
\alias{fetchKey}
\title{Fetch API Key}
\usage{
fetchKey(token)
}
\arguments{
\item{token}{Token obtained with \code{\link{signIn}}}
}
\value{
A character containing your API key.
}
\description{
Fetch API Key
}
\examples{
token <- "123456789"
fetchKey(token)
}
