% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateKeys.R
\name{generateKeys}
\alias{generateKeys}
\title{Generate new API key & API secret}
\usage{
generateKeys(token, as_df = TRUE)
}
\arguments{
\item{token}{Token obtained with \code{\link{signIn}}}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with the new API key and API secret key.
}
\description{
With this function you can generate new pair of API Key & API Secret.
Every time you generate a new pair, an old one becomes invalid.
Your API Secret won’t be accessible later, so please note it down in some secret place.
If you have lost your API Secret, you can generate a new pair of API Key & API Secret here.
}
\examples{
token <- "123456789"
generateKeys(token, as_df = TRUE)
}
