% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{integrations}
\alias{integrations}
\title{List enabled Integrations}
\usage{
integrations(token, as_df = TRUE)
}
\arguments{
\item{token}{Token obtained with \code{\link{signIn}}}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with names of the integrations.
}
\description{
With this endpoint you can list names of all Integrations you have enabled.
}
\examples{
token <- "123456789"
integrations(token, as_df = TRUE)
}
