% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listActivities.R
\name{listActivities}
\alias{listActivities}
\title{List all activities}
\usage{
listActivities(token, as_df = TRUE)
}
\arguments{
\item{token}{Token obtained with \code{\link{signIn}}}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with id, name, color, integration and device side
of the activities.
}
\description{
List all activities
}
\examples{
token <- "123456789"
listActivities(token, as_df = TRUE)
}
