% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listArchivedActivities.R
\name{listArchivedActivities}
\alias{listArchivedActivities}
\title{List all archived activities}
\usage{
listArchivedActivities(token, as_df = TRUE)
}
\arguments{
\item{token}{Token obtained with \code{\link{signIn}}.}

\item{as_df}{If \code{TRUE} a data frame is returned, if \code{FALSE} a list is returned.}
}
\value{
A data frame or list with id, name, color, integration and deviceSide (\code{NULL}).
}
\description{
List all archived activities
}
\examples{
token <- "123456789"
listArchivedActivities(token, as_df = TRUE)
}
