% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDevices.R
\name{listDevices}
\alias{listDevices}
\title{List all known devices}
\usage{
listDevices(token, as_df = TRUE)
}
\arguments{
\item{token}{Token obtained with \code{\link{signIn}}}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with serial number, name, active and disabled indicators.
}
\description{
List all known devices
}
\examples{
token <- "123456789"
listDevices(token, as_df = TRUE)
}
