% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posix_to_timeular_timestamp.R
\name{posix_to_timeular_timestamp}
\alias{posix_to_timeular_timestamp}
\title{Convert POSIX object to Timeular timestamp}
\usage{
posix_to_timeular_timestamp(posix = "2017-09-24 17:18:53 CEST")
}
\arguments{
\item{posix}{POSIX object}
}
\value{
A timestamp with Timeular style
}
\description{
Convert POSIX object to Timeular timestamp
}
\examples{
posix_to_timeular_timestamp(posix = "2017-09-24 17:18:53 CEST")
}
