% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagsAndMentions.R
\name{tagsAndMentions}
\alias{tagsAndMentions}
\title{Fetch tags & mentions of given activity}
\usage{
tagsAndMentions(activityId = NULL, token, as_df = TRUE)
}
\arguments{
\item{activityId}{ID of an Activity, eg. 123.}

\item{token}{Token obtained with \code{\link{signIn}}.}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with type (tag, mention or error), id and label.
}
\description{
Tags and mentions are created with use of # and @ prefixes in notes of your
time entries. Moreover if an activity is linked with integration, let’s say
JIRA project, JIRA task IDs are visible as tags. With this endpoint you can
fetch all tags and mentions valid in context of given activity. In this API
version each tag / mention has ID only, while labels are \code{NULL} / \code{NA}.
}
\examples{
token <- "123456789"
tagsAndMentions(activityId = "57302", token, as_df = TRUE)
}
