% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetrackDay.R
\name{timetrackDay}
\alias{timetrackDay}
\title{Get the aggregated timetrack on a specific day}
\usage{
timetrackDay(day = as.POSIXct("2017-09-21", tz = "CET"), token, tz = "CET")
}
\arguments{
\item{day}{A POSIX object of the day one wishes to get time tracking from}

\item{token}{Token obtained with \code{\link{signIn}}}

\item{tz}{The desired timezone}
}
\value{
A data frame with activity name (\code{name}), time difference in minutes (\code{timeDiff}),
time difference represented as H:M (\code{hm}) and the time difference represented as decimal hours (\code{decHour}).
}
\description{
Get the aggregated timetrack on a specific day
}
\examples{
token <- "123456789"
day <- as.POSIXct("2017-09-21", tz = "CET")
timetrackDay(day, token, tz = "CET")
}
