% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetrackWeek.R
\name{timetrackWeek}
\alias{timetrackWeek}
\title{Get the aggregated timetrack in a specific week}
\usage{
timetrackWeek(weekNumber = 38, token, year = NULL, tz = "CET")
}
\arguments{
\item{weekNumber}{The week number you wish timetracking from}

\item{token}{Token obtained with \code{\link{signIn}}}

\item{year}{If \code{NULL} the current year is choosen}

\item{tz}{The desired timezone}
}
\value{
A data frame containing the date \code{date}, week day \code{weekDay},
project name \code{name}, time difference represented as H:M (\code{hm}) and
the time difference represented as decimal hours (\code{decHour}).
}
\description{
Get the aggregated timetrack in a specific week
}
\examples{
token <- "123456789"
timetrackWeek(weekNumber = 38, token, year = NULL, tz = "CET")
}
