% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unassignActivity.R
\name{unassignActivity}
\alias{unassignActivity}
\title{Unassign an activity from a device Side}
\usage{
unassignActivity(activityId = NULL, deviceSide = NULL, token,
  as_df = TRUE)
}
\arguments{
\item{activityId}{ID of an Activity, eg. 123.}

\item{deviceSide}{Side of an active Device.}

\item{token}{Token obtained with \code{\link{signIn}}.}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with id, name, color, integration and deviceSide (\code{NULL}).
}
\description{
With this endpoint you can unassign an activity from side of your active Device.
}
\examples{
token <- "123456789"
unassignActivity(activityId = "57302", deviceSide = 8, token, as_df = TRUE)
}
