% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userProfile.R
\name{userProfile}
\alias{userProfile}
\title{Fetch user's profile}
\usage{
userProfile(token, as_df = TRUE)
}
\arguments{
\item{token}{Token obtained with \code{\link{signIn}}}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with userId, email, firstName and lastName.
}
\description{
Fetch user's profile
}
\examples{
token <- "123456789"
userProfile(token, as_df = TRUE)
}
