% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserProfile.R
\name{UserProfile}
\alias{UserProfile}
\title{Fetch user's profile}
\usage{
UserProfile(token, as_df = TRUE)
}
\arguments{
\item{token}{Token obtained with \code{\link{DeveloperSignInRequest}}}

\item{as_df}{If TRUE a data frame is returned, if FALSE a list is returned.}
}
\value{
A data frame or list with userId, email, firstName and lastName.
}
\description{
Fetch user's profile
}
\examples{
token <- "123456789"
UserProfile(token, as_df = TRUE)
}
